Explicitly modprobe ext2 before mkinitrd
mount on RHEL 3 isn't clever about ensuring the appropriate module is loaded for
the filesystem it's trying to mount. mkinitrd on RHEL 3 uses mount, and fails
because ext2 isn't loaded. To work round this, we explicitly load ext2 before
running mkinitrd. This makes RHEL 3 work, and doesn't hurt anything else.