virt-snapshot: Write domain to temp file on error
Write the domain xml to a temporary file and return an error if we weren't able
to create a domain and --outputxml wasn't specified. The error message includes
the name of the temporary file. This avoids potential data loss in --rollback if
we've already deleted the storage but can't restore the old domain.