v2v-snapshot: Check volumes are snapshot volumes before deleting them
This goes part-way to addressing RHBZ 524268. If v2v-snapshot is run against a
live domain, the storage changes will not be effective until the domain is shut
down. Running v2v-snapshot --rollback before the domain is shut down will find
the original storage. It deletes it without checking it is a snapshot volume.

This change ensures that a volume is a snapshot volume before deleting it. It
stops processessing if it is not.