Converter: Prefer detected arch of OS if it's available
A 32bit VMware guest will boot a 64bit OS. However, a 32bit KVM guest won't.
This means that if you accidentally selected a 32bit machine type at VM creation
time on ESX, then installed a 64bit OS, chances are you won't notice until you
convert it and it doesn't work any more. To get round this, we always use the
detected architecture of the OS if it's available. This will mean that your
32bit guest will become a 64bit guest if it has a 64bit OS on it.