Try hard to clean up on unclean shutdown
Set autosync on the libguestfs handle, install signal handlers for SIGINT and
SIGQUIT which explicitly close the handle, and add an END { } block which closes
the handle in virt-v2v.pl.

Note that these measures still don't work when the user kills virt-v2v with
Ctrl-C. We suppose this is because qemu is also receiving the SIGINT and
shutting down before it can be shut down cleanly by libguestfs. However, the
above changes should improve matters if this can be fixed.