Fix remapping of block devices
We were assuming that guests wouldn't use multiple interfaces, or if they did,
drive letters wouldn't clash between the interfaces. An ESX guest with a CDROM
device breaks this because hard disks will be presented a SCSI, starting at sda,
and CDROM devices are presented as IDE, starting at hda. This is in contrast to
QEMU, which starts at hdc by default.

Firstly, this change causes disks and CDROM devices to be treated differently.
CDROM devices are changed to use IDE, regardless of what interface they were
using before. This also limits the maximum number of CDROM devices in a guest to
4. CDROM devices after 4 will be dropped, and a warning emitted.

Secondly, it is smarter about renaming IDE and SCSI devices, which may be merged
in the guest configuration.

Thirdly, it renames devices consistently based on the order they are listed in
the domain XML.