v2v: Fix error on exit unmounting transfer ISO
If the transfer iso was mounted during conversion, virt-v2v would always give
the following error on shutdown:

    (in cleanup) umount: /tmp/transferb7icam: umount:
    /sysroot/tmp/transferb7icam: not found at
    /home/mbooth/src/virt-v2v/blib/lib/Sys/VirtV2V/GuestOS/RedHat.pm line 1171.

This was because the GuestOS::RedHat cleanup was being called implicitly on
exit, which is after umount_all has been called explicitly on the libguestfs
handle.  This change garbage collects the guestos explicitly before the
libguestfs handle.