Workaround issue where grubby fails when run under libguestfs
When installing kernel-2.6.18-128.el5 in a guest with
kernel-xen-2.6.18-164.el5xen already installed, and 'root=LABEL=/' on the kernel
command line, grubby fails to install a grub entry for the new kernel. It only
fails when run under libguestfs. Installing in the original Xen guest works
fine. The underlying cause appears to be related to blkid, although I haven't
managed to pin it down exactly.

This change works around the issue. If virt-v2v encounters a problem with grub,
it will install a new kernel entry itself using augeas.

Note that this change introduces a dependency on the new aug_clear libguestfs
command.