ESX: Get volume size in a separate HEAD request
We used to fetch the volume in a HEAD request. After receiving the header, which
contains the content length, we created the target volume and opened it for
writing before continuing to receive data. This created a problem if volume
creation took long enough for the transfer to time out.

To fix this, we get the size first with a HEAD request. Create the volume, then
do a second GET request to receive the data.

This fixes RHBZ#586816