SSH: Don't mix buffered reads with sysread
SSH transfer contained a race due to use of buffered IO to read the initial
size, followed by sysread() to read the data. If data was available for reading
at the time the size was read, some of it would be consumed and buffered. It
would subsequently not be available to sysread(), meaning it wasn't copied.

Fix by using buffered IO consistently.