GuestOS: Fix augeas grub configuration
Augeas now configures the Grub lens for /boot/grub/menu.lst by default. The code
which was checking this was broken in the case where there was more than 1
included path.

We now check if /boot/grub/menu.lst is included. If it is, we don't reload. If
it isn't, we add it, but don't remove any other config first.

This fixes RHBZ#586822