Config: Don't require all referenced software to be available
Currently virt-v2v attempts to create an iso containing all software referenced
in the config file before starting. If any software isn't available it displays
an error and exits immediately.

With this change it will ignore software which isn't available at iso creation
time. Instead, it will only display an error if the unavailable software is
actually required. This is much kinder to the casual user using the default
configuration file.