Config: Add --bridge and --network command line options
This change allows a default network mapping to be specified on the command line
which will be used if no mapping was found in the configuration file, or if no
configuration file was specified.

This change makes it possible to convert all guests except Xen PV guests without
a config file. Xen PV guests will require a config file to specify replacement
kernels. Guests with an old kernel that don't support VirtIO still need the
config file to enable VirtIO support.