List all missing dependencies at once
When a conversion requires a file from the config file which isn't present, we
currently die and report which file was missing. If there are many missing
dependencies, using this method to find them all would take multiple runs.

This change attempts to find all missing dependencies in the first run.