Fix import of RHEL 3 kvm guests using kmod-virtio
RHEL 3 doesn't have a dynamic /dev. kmod-virtio creates devices nodes for vdX
block devices based on what the major number of virtblk was at the time it was
installed. This is, in turn, based on the order the modules were loaded in
initrd.

To try to preserve this precarious state of affairs, when adding virtio drivers
to a new initrd we always load the same drivers as kmod-virtio in the same
order.