Add capabilities
Before RHN support was added, the only way to install software was by specifying
it in the config file. To configure VirtIO support, Converter::Linux called
add_kernel(). We setup the config file such that the kernel version in there
supported VirtIO, and everything else which needed to be installed was a
dependency of the kernel. This meant that the RHN support had no good way to
determine from configuration what needed to be installed, and at what versions.

Capabilities describe a set of requirements for a feature, which means they can
be used cleanly for both RHN support, and installation from local sources. They
will also allow VirtIO to be configured in RHEL 3 guests by installing the
kmod-virtio package rather than a new kernel, and are intended to be usable for
installing spice.

This change principally adds install_capability to GuestOS::RedHat, replacing
add_kernel. It also adds a new function, install_good_kernel, which will
specifically attempt to install a bootable kernel from any source. It also
refactors the RHN support to make it usable by both these functions, and adds
additional error checking.