ESX: Fix storage URL if storage has a snapshot
If an ESX guest has a snapshot, the path the libvirt driver gives us will look
like:
  [yellow:storage1] RHEL4-X/RHEL4-X-000003.vmdk
instead of:
  [yellow:storage1] RHEL4-X/RHEL4-X.vmdk
The current path mangling code does take this into account.

This change makes it try the original name first, but try again after removing a
'-\d+' suffix if it gets a 404. Trying twice should make it continue to work in
the event that a local naming scheme matches the suffix used for snapshots.
Ideally we would be able to get this information from the libvirt driver, but
it's not implemented yet.

Fixes RHBZ#600220