Fix cleanup if Ctrl-C kills guestfs qemu process
If a user kills v2v with Ctrl-C during the conversion stage, there is a high
likelihood that the qemu process will die before v2v cleanup completes. When
this happens, the unmount_all and sync calls will fail. This causes all further
cleanup to be aborted, which is undesirable.

Additional fix to RHBZ#596015