Correctly detect the size of a block device over SSH
When fetching a storage volume over SSH, we were detecting its size with a
simple 'stat -c %s'. While this works fine for files, it won't return the
correct size of a block device.

This patch uses the output of 'blockdev --getsize64' for block devices, and stat
as before for regular files.

Fixes RHBZ#620698