Fix conversion of guest with multiple cdrom devices in boot order
One of my Xen guests has multiple cdrom devices in its domain XML. I don't know
how this happened, but it works. However, QEMU will not allow this, and fails to
start.

This patch checks for multiple devices of the same type, and removes all but the
first instance of a device of a particular type.