Check that a RHEV export storage domain has been attached to a Data Center
The master/vms directory in an export storage domain is only created when the
domain is attached to a Data Center, not when it is initialised. Because we
write to this directory last, virt-v2v will only currently fail at the very end
of the conversion process. This patch checks for the existence of this directory
early and reports a useful error to the user.

Fixes RHBZ#601535