Only check total data received when transferring raw volumes
When transferring a qcow2 image, the total data received doesn't correspond
exactly to either the allocated or total size of image. In my test, it was
actually allocated + 4k, which I assume accounts for header information. Until I
find documentation on exactly what is expected, just skip this test for non-raw
images.