Don't use helper processes for NFS in RHEVTarget
For every operation which needed to read or write to NFS in RHEVTarget, we were
previously spawning a new process which ran set(u|g)id 36:36. We used a
different technique, however, to spawn the conversion appliance when its storage
was hosted on RHEV: sete(u|g)id before and after, and push all error handling
outside the unprivileged code.

This change updates RHEVTarget to use the latter technique instead of helper
processes. The primary driver for doing this is that it gives almost a 100%
throughput improvement when writing to a guestfs appliance during format
conversion.