Remove GuestOS
Converter::Linux becomes Converter::RedHat, and everything from GuestOS::RedHat
is folded into it.

GuestOS was intended to be an OS-specific interface for a generic converter.
This turned out to be impractical, and it was relegated to an interface for the
Linux converter. In the absence of support for any non-Red Hat-based
distribution, it is doubtful the existing interface would have been useful. It
also made the code significantly harder to follow. The intention is to refactor
Converter::RedHat again to enable re-use of any generally useful code when we
eventually add support for other distributions.