Fix architecture detection of Linux guests with invalid grub.conf
If a Linux guest's grub.conf contained a default kernel which referred to
non-existent kernel, architecture detection would default to i686. For an x86_64
guest, the result would not boot.

This change uses the first grub.conf entry if the default is invalid or not
present, and falls back to the architecture detected by inspection otherwise. If
neither of these yield an architecture, it defaults to x86_64 instead of i686 as
a more conservative default.

Fixes RHBZ#623579