Fix detection of an existing grub entry
There were 2 issues in the code which checked for an existing grub entry before
adding a new one.

Firstly, it didn't take account of the fact that the passed-in kernel path is
relative to root, whereas the grub entry is relative to the grub filesystem,
which is normally /boot.

Secondly, it expected return from inside eval{} to exit the function, when in
fact it only exits the eval.