Don't remove foreign kernels during conversion
Removing xen kernels during conversion is tidier, but occasionally problematic.
For example, if the kernel has dependent kmods which prevent its uninstallation,
the code to remove it reliably across multiple versions of RHEL becomes tricky.
The existing code doesn't handle this at all, and the conversion will fail in
this case.

As it isn't strictly necessary to remove xen kernels, this change simply stops
attempting to remove them.

Fixes RHBZ#643867