Always use user_message for die() messages in Config
I removed a bunch of user_message calls in Config when calling die() whilst
refactoring it. The idea was to catch the die() higher up the stack and wrap the
message there, which would be cleaner. Unfortunately this doesn't work, as
without a newline in the argument to die(), perl adds source context information
to the message immediately, rather than when the message is ultimately
displayed. This means that user_message just wraps a message that already has
source context information in it, which is not the desired effect. This change
puts the calls back.

It also makes an NFC stylistic change to all other die() calls in Config by
removing the parentheses. I've decided that die and warn are cleaner without.