Display verbose error message when guest storage isn't found
Unfortunately libvirt will not always return volume information for a guest's
volume. Specifically, it will not return volume information if the volume is not
in a storage pool. Rather than displaying a generic libvirt error in this case,
advise the user how they can create an appropriate storage pool.

Fixes RHBZ#672498