Replace File::Path's remove_tree
We were using remove_tree in the RHEV cleanup code. remove_tree seems to use
chdir. Unfortunately, as this code will be running seteuid(36), it is not
unlikely that the current working directory will not be readable. This causes
remove_tree to fail.

This patch replaces remove_tree with a simple rm -rf. We also only warn on
failure, as it's not strictly fatal.

Fixes RHBZ#670778