Check for, and uninstall, VMware Tools installed from tarball
VMware Tools can be installed by RPM or from a tarball. We were only looking for
the RPM. If it was installed by tarball, we would end up leaving it in place. If
VMware Tools detects that it is not running on a VMware platform when it starts,
it will clumsily try to uninstall itself, leading to misconfiguration.

This patch looks for tarball-installed VMware Tools and uninstalls it.

Fixes RHBZ#623571