Allow conversion of Windows guests with partial VirtIO support
There are currently no netkvm drivers available for Windows XP or Windows 7.
Currently virt-v2v will not detect this, and will always enable VirtIO
networking. This change detects that there is no netkvm driver, and instead
configures an rtl8139 driver. The rtl8139 driver is chosen over the Linux
default of e1000 as it is more likely to be already installed. This
required changing guestcaps to contain explicit block and net drivers, rather
than a single virtio boolean.

This change also consequently enables conversion of Windows guests with no
VirtIO drivers available. However, it makes no attempt to install either ide or
rtl8139 drivers in the guest. If they are not available in the guest, it will
not operate correctly. In the case of ide, it will not boot. A warning message
is displayed in this case.

Fixes RHBZ#676553