Allow reading more data than the reported size of a volume
If a volume is not an exact multiple of 512 bytes, qemu-img will report its size
rounded down to a 512 byte boundary. However, when copying, the file is still
read until EOF, which will return more data than was expected. This change
prevents that causing a failure in itself.

The situation is still not resolved, however, as there are still situations
where this will cause a failure. For example, copying will fail writing to a
block device if an attempt is made to write too much data.