Add --root (root choice) option.
This option allows the root filesystem to be chosen more
intelligently for operating systems that are (correctly or
otherwise) identified as multi-boot.

This changes the default behaviour so that instead of just
printing an error and dying, which all users hated, it now
defaults to asking the user which root filesystem they want
to use.  You can get the old behaviour by specifying
--root=single.  There are various other behaviours available
too.