p2v: Add virt-p2v-launcher
virt-p2v-launcher is a wrapper which calls virt-p2v and checks that it completed
successfully. If it didn't, it prompts the user to try again, debug or power
off.

This change also checks that X launched successfully, and gives the user the
same options as virt-p2v-launcher if it didn't.