build: Split v2v and p2v into separate spec files
v2v depends on libguestfs, which depends on KVM, which is only available on
certain architectures. In RHEL, it's only available on x86_64. Due to
limitations of koji, we must therefore make v2v arch-dependent to avoid it being
built on a platform where libguestfs/kvm isn't available.

p2v must be i386 to allow a single boot image to work on both i386 and x86_64.
Because we now have 2 different architecture requirements, we have to split the
spec file.