Make transfer iso mounting/unmounting explicit
The transfer iso is currently mounted implicitly the first time it is required,
and subsequently never expected to change mount point. However, if the root disk
is unmounted and replaced with a new root disk, the mount point created for the
transfer iso will not be cleaned up from the old disk, and will not be present
on the new disk.

This change makes mounting and unmounting explicit. This is required for
multiboot support.