Use a temporary file location for transfer iso
Now that we regenerate the transfer iso on every run, there is no reason to
require the user to specify where it goes. This change makes is simpler to run
virt-v2v as non-root, as the user doesn't have to override the default config to
specify a writable transfer iso location.