Fix changing allocation policy of qcow2 storage
When converting a guest with qcow2 storage and changing the allocation policy
(preallocated to sparse or vice versa), virt-v2v would simply copy the existing
qcow2 file without modifying it, but the file created would be either
preallocated or sparse. This change causes virt-v2v to create a new qcow2 file
with the correct allocation policy.

In the same change, it also enables format conversion of non-raw storage over
SSH.