Don't lose VNC password during conversion
If v2v was converting a libvirt guest which required a VNC password, the
converted libvirt guest would not require a VNC password. This change ensures we
don't lose the password during libvirt->libvirt conversion. RHEV doesn't have
anything directly analogous to this, so when converting libvirt->RHEV we display
a warning if the source guest required a VNC password.

Fixes CVE-2011-1773