p2v-client: Don't use AIO while reading
I originally added AIO as an attempt to fix a performance issue which it turns
out was caused by something else. I left it in because it worked and there was
no reason to remove it. I have recently hit a hard to debug crash in the P2V
client which is complicated by, if not caused by, the additional AIO thread.

This change removes AIO in favour of synchronous reads to simplify the code.