p2v-server: Don't die on libguestfs warning messages
The P2V client will die if it receives anything on stderr. Both qemu and current
libguestfs write warning messages to stderr, which will cause the client to die
unnecessarily. This change redirects anything written to stderr to a log file,
which is cleaned up on exit if nothing was written to it.