Don't preallocate qcow2 for RHEV targets
When creating a sparse qcow2 image for output to a RHEV guest, we were
preallocating the metadata in advance. This caused problems in RHEV, which
didn't recognise that the created file was sparse. Testing suggests that the
qemu performance problem which prompted preallocating metadata has now been
fixed, so we're removing it.