Fix mapping of PV block devices in Xen HVM guests
Xen guests may specificy PV or non-PV block devices. If a device is specified as
non-PV, Xen also presents a PV device with the same backing store. This means we
need to be prepared to see PV devices in Xen guests, even if the metadata
indicates they are not there.

Fixes RHBZ#737855