RedHat: Check grub after installing a kernel
This fixes a regression which caused a RHEL 5 conversion to fail when installing
a new virtio kernel. We were only checking that the grub entry existed at the
very end of the conversion process, but as list_kernels enumerates based on
grub it was failing to enumerate the newly installed kernel.